+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2026-02-01
+
+### Fixed
+
+- Client address fields (street, city, state, postcode) now save correctly to ServiceM8
+  - Fixed parsing of n8n fixedCollection data structure for address inputs
+
 ## [1.0.0] - 2026-02-01
 
 ### Added